projects/crypto-liquidity-hunter/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
ccxt>=4.0.0
ta>=0.10.0
sqlalchemy>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
flask>=3.0.0
plotly>=5.14.0
dash>=2.14.0
psutil>=5.9.0
pytz>=2022.7
PyYAML
TA-Lib # optional, requires system install
